Handle relative path arguments in Killfilter
KillFilter currently expects an absolute path
to be specified for the process name to kill. This
is inconvenient when the exact location of the running
binary is not known or differs accross installs.

Extend KillFilter to accept also commands in $PATH
to be killed if the given argument is not absolute.

Change-Id: I6b90206b587ff3f949af2c256a78ca21af31867a