Trivial changes from oslo-incubator
A number of trivial changes were merged to the oslo-incubator copy of
the rootwrap code while the code was being graduated. Those changes
should have been rejected (as the code was frozen while in graduation)
and applied to oslo.rootwrap instead.

Now that the oslo-incubator copy of the code was removed, we should
reapply those trivial changes to oslo.rootwrap instead, otherwise they
will just be lost:

* Remove vim header (original author: jogo)
* Make vertical white space after license header consistent (noorul)
* Removed copyright from empty files (jay-lau)

Change-Id: I0403335b8ec151e2337d50935be18355cc55a199