Fix portability issue
The use of the "resource" module now prevents oslo.rootwrap from
being imported on Windows.

Although oslo.rootwrap is not effectively used on Windows, it's
important for it to at least be importable, since it's widely
used throughout OpenStack projects without having platform checks
in place.

This change checks if the "resource" module is avaialble before
attempting to use it.

Change-Id: I2391315f77718a3c9eb9fc8c03a6882237f33548
Closes-Bug: #1804639