[stable-only] Cap bandit and fix lower-constraints
The 1.6.3 [1] release has dropped support for py2 [2] but the release
is faulty and pip still picks it up for py2 [3][4], so cap to 1.6.2
when using py2.

Contradicting hacking version replaced (in lower-constraints.txt to
match with test-requirements.txt), which pulls in newer flake8, too.

Sphinx requirements also fixed to make requirements-check job pass.

[1] https://github.com/PyCQA/bandit/releases/tag/1.6.3
[2] https://github.com/PyCQA/bandit/pull/615
[3] https://github.com/PyCQA/bandit/issues/663
[4] https://github.com/PyCQA/bandit/issues/665

Change-Id: Ibdc46d188a6025a42a6c7244cb0700563f56b3d1