Add coreutils as valid value to fix KillFilter test
In containerized RHEL 9 environment the KillFilter test started to fail
because `os.readlink("/proc/<pid-of-cat/exe>")` return /usr/bin/coreutils.

Indeed cat is part of the coreutils package and the symlink of
this file point toward coreutils. Fixing this bug by adding
coreutils as a possible returned value for our tests.

Closes-Bug: #2037383
Change-Id: I141eb15efa30c7df3ca5419b594097f4b683663c